Rooms CRUD (#15)
* feat(modules): create rooms module and initial controller

* feat(rooms): create room entity and DTOs

* feat(rooms): create service

* feat(app): replace getHello with getWelcome

* feat(rooms): move files to rooms folder

* feat(rooms): create CRUD

* feat(rooms): relationate with floors

* fix(rooms): add ParseIntPipe in controller

* feat(rooms): use ApiResponse in controller

* fix(floors,rooms): unit testings

* feat(rooms): test controller and service

* fix(room-category): add controller and service

* fix(dependencies): re-install all dependencies

+++ src/rooms/controllers/rooms/rooms.controller.ts
@@ -9,32 +9,49 @@ import {
   Body,
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
+
+import { RoomsService } from '../../services/rooms/rooms.service';
+import { CreateRoomDto, UpdateRoomDto } from '../../dtos/rooms.dto';
 import {
   CreateRoomCategoryDto,
   UpdateRoomCategoryDto,
 } from '../../dtos/rooms-categories.dto';
-
-import { CreateRoomDto, UpdateRoomDto } from '../../dtos/rooms.dto';
-import { RoomsService } from '../../services/rooms/rooms.service';
+import { ApiResponse } from '../../../common/api/apiResponse';
+import { Code } from '../../../common/code/code';
 
 @ApiTags('Rooms')
 @Controller('rooms')
 export class RoomsController {
   constructor(private roomsService: RoomsService) {}
 
-  @Get()
-  getRooms() {
-    return this.roomsService.findAll();
+  @Post()
+  async createRoom(@Body() payload: CreateRoomDto) {
+    const roomAlreadyExists = await this.roomsService.findOneBy({
+      number: payload.number,
+    });
+    if (roomAlreadyExists)
+      return ApiResponse.error(
+        Code.ENTITY_ALREADY_EXISTS_ERROR.code,
+        Code.ENTITY_ALREADY_EXISTS_ERROR.message,
+      );
+    const newRoom = await this.roomsService.create(payload);
+    return ApiResponse.created(newRoom);
   }
 
-  @Post()
-  createRoom(@Body() payload: CreateRoomDto) {
-    return this.roomsService.create(payload);
+  @Post('/categories')
+  createCategory(@Body() payload: CreateRoomCategoryDto) {
+    return this.roomsService.createCategory(payload);
+  }
+
+  @Get()
+  async getRooms() {
+    const rooms = await this.roomsService.findAll();
+    return ApiResponse.success(rooms);
   }
 
   @Get('/categories')
-  getCategories() {
-    return this.roomsService.findAllCategories();
+  async getCategories() {
+    return await this.roomsService.findAllCategories();
   }
 
   @Get('/categories/:id')
@@ -42,9 +59,15 @@ export class RoomsController {
     return this.roomsService.findOneCategory(id);
   }
 
-  @Post('/categories')
-  createCategory(@Body() payload: CreateRoomCategoryDto) {
-    return this.roomsService.createCategory(payload);
+  @Get('/:id')
+  async getRoom(@Param('id', ParseIntPipe) id: number) {
+    const room = await this.roomsService.findOne(id);
+    if (!room)
+      return ApiResponse.error(
+        Code.NOT_FOUND_ERROR.code,
+        Code.NOT_FOUND_ERROR.message,
+      );
+    return ApiResponse.success(room);
   }
 
   @Put('/categories/:id')
@@ -55,26 +78,35 @@ export class RoomsController {
     return this.roomsService.updateCategory(id, payload);
   }
 
+  @Put('/:id')
+  async updateRoom(
+    @Param('id', ParseIntPipe) id: number,
+    @Body() payload: UpdateRoomDto,
+  ) {
+    const room = await this.roomsService.findOne(id);
+    if (!room)
+      return ApiResponse.error(
+        Code.NOT_FOUND_ERROR.code,
+        Code.NOT_FOUND_ERROR.message,
+      );
+    const updatedRoom = await this.roomsService.update(room, payload);
+    return ApiResponse.success(updatedRoom);
+  }
+
   @Delete('/categories/:id')
   deleteCategory(@Param('id', ParseIntPipe) id: number) {
     return this.roomsService.deleteCategory(id);
   }
 
-  @Get('/:id')
-  getRoom(@Param('id', ParseIntPipe) id: number) {
-    return this.roomsService.findOne(id);
-  }
-
-  @Put('/:id')
-  updateRoom(
-    @Param('id', ParseIntPipe) id: number,
-    @Body() payload: UpdateRoomDto,
-  ) {
-    return this.roomsService.update(id, payload);
-  }
-
   @Delete('/:id')
-  deleteRoom(@Param('id', ParseIntPipe) id: number) {
-    return this.roomsService.delete(id);
+  async deleteRoom(@Param('id', ParseIntPipe) id: number) {
+    const room = await this.roomsService.findOne(id);
+    if (!room)
+      return ApiResponse.error(
+        Code.NOT_FOUND_ERROR.code,
+        Code.NOT_FOUND_ERROR.message,
+      );
+    await this.roomsService.delete(id);
+    return ApiResponse.success(room);
   }
 }
